Update dependencies and enhance error handling in CLI and main modules
- Added new dependencies: `torch`, `pydantic`, `pyyaml`, and `tomli` for improved functionality and compatibility.
- Enhanced error handling in `caramba/__main__.py` to provide clearer messages for various exceptions.
- Updated `CLI` class in `caramba/cli.py` to improve command parsing and error reporting.
- Refactored layer and topology builders to ensure better type safety and error handling.
- Included tests for new functionality and error cases to ensure robustness.

+++ caramba/operation/attention_math.py
@@ -18,7 +18,7 @@ def shape_heads(x: Tensor, *, n_heads: int, head_dim: int) -> Tensor:
         raise ValueError(f"n_heads must be > 0, got {n_heads}")
     if head_dim <= 0:
         raise ValueError(f"head_dim must be > 0, got {head_dim}")
-    if int(x.shape[-1]) != int(n_heads * head_dim):
+    if x.shape[-1] != n_heads * head_dim:
         raise ValueError(
             "Expected last dim to equal n_heads*head_dim, got "
             f"x={x.shape}, n_heads={n_heads}, head_dim={head_dim}"
